Make sure that all qlog texts go through JSON filter

+++ picoquictest/ticket_store_test.c
@@ -571,7 +571,7 @@ int token_reuse_cap_test(void)
         for (size_t i = 0; ret == 0 && i < nb_tokens; i++) {
             picoformat_64(token + 4, (uint64_t)i);
             if (picoquic_registered_token_check_reuse(quic, token, sizeof(token), (uint64_t)i) != 0 &&
-                quic->token_reuse_tree.size < token_reuse_cap) {
+                (size_t)quic->token_reuse_tree.size < token_reuse_cap) {
                 DBG_PRINTF("Token[%z] unexpectedly flagged as reused", i);
                 ret = -1;
             }
